update-code-remove-ai-changed-rules
