+++ Tic-Tac-Toe frontend/tic-tac-toe/src/App.js
@@ -2,7 +2,7 @@
 import './App.css';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import HomePage from './pages/HomePage';
-import PlayWithAI from './pages/PlayWithAI';
+import Play from './pages/Play';
 import Login from './pages/Login';
 import Register from './pages/Register';
 import PlayOnline from './pages/PlayOnline';
@@ -14,7 +14,7 @@ function App() {
     <Route path='/' element={<HomePage />}/>
     <Route path='/login' element={<Login />}/>
     <Route path='/register' element={<Register />}/>
-    <Route path="/playai" element={<PlayWithAI />}/>
+    <Route path="/playai" element={<Play />}/>
     <Route path="/playonline" element={<PlayOnline />}/>
 
    </Routes>
